+++ backend/be-digital-booking/src/main/java/com/grupo9/db/service/ProductService.java
@@ -23,36 +23,46 @@ public class ProductService {
         this.responsesBuilder = responsesBuilder;
     }
 
-    public ResponseEntity<ApiResponse> findAll(){
+    public ResponseEntity<ApiResponse<List<Product>, Object>> findAll(){
         List<Product> products = repository.findAll();
-        return responsesBuilder.buildResponse(HttpStatus.OK.value(),"Get Product List successfully",products);
+        return responsesBuilder.buildResponse(HttpStatus.OK.value(),"Get Product List successfully",products, null);
     }
 
-    public ResponseEntity<ApiResponse> findById(Long id) throws ResourceNotFoundException, BadRequestException {
-        if(id == null) throw new BadRequestException("ID missing");
+    public ResponseEntity<ApiResponse<Product, Object>> findById(Long id) throws ResourceNotFoundException {
         Optional<Product> product = repository.findById(id);
         if(product.isEmpty()){
             throw new ResourceNotFoundException("Product with id " + id + " not found");
         }
-        return responsesBuilder.buildResponse(HttpStatus.OK.value(),"Get Product successfully", product.get());
+        return responsesBuilder.buildResponse(HttpStatus.OK.value(),"Get Product successfully", product.get(), null);
     }
 
-    public ResponseEntity<ApiResponse> save(Product product){
+    public ResponseEntity<ApiResponse<Product, Object>> save(Product product){
         Product response = repository.save(product);
-        return responsesBuilder.buildResponse(HttpStatus.CREATED.value(),"Product created successfully", response);
+        return responsesBuilder.buildResponse(HttpStatus.CREATED.value(),"Product created successfully", response, null);
     }
 
-    public ResponseEntity<ApiResponse> update(Long id, Product product) throws ResourceNotFoundException, BadRequestException {
+    public ResponseEntity<ApiResponse<Product, Object>> update(Long id, Product product) throws ResourceNotFoundException, BadRequestException {
+        if(id == null) throw new BadRequestException("ID missing");
+
+        Boolean exists = repository.existsById(id);
+        if(!exists){
+            throw new ResourceNotFoundException("Product with id " + id + " not found");
+        }
+
         product.setId(id);
-        this.findById(product.getId());
         Product response = repository.save(product);
-        return responsesBuilder.buildResponse(HttpStatus.CREATED.value(),"Product updated successfully", response);
+        return responsesBuilder.buildResponse(HttpStatus.CREATED.value(),"Product updated successfully", response, null);
     }
 
     public ResponseEntity<ApiResponse> deleteById(Long id) throws ResourceNotFoundException, BadRequestException {
-        this.findById(id);
+        if(id == null) throw new BadRequestException("ID missing");
+        Boolean exists = repository.existsById(id);
+        if(!exists){
+            throw new ResourceNotFoundException("Product with id " + id + " not found");
+        }
+
         repository.deleteById(id);
-        return responsesBuilder.buildResponse(HttpStatus.NO_CONTENT.value(),"Product deleted successfully", "");
+        return responsesBuilder.buildResponse(HttpStatus.OK.value(),"Product deleted successfully", null, null);
     }
 
 }
